+++ scm/driver/gitea/repo.go
@@ -46,6 +46,10 @@ func (s *repositoryService) List(ctx context.Context, opts scm.ListOptions) ([]*
 	return convertRepositoryList(out), res, err
 }
 
+func (s *repositoryService) List2(ctx context.Context, orgSlug string, opts scm.ListOptions) ([]*scm.Repository, *scm.Response, error) {
+	return nil, nil, scm.ErrNotSupported
+}
+
 func (s *repositoryService) ListHooks(ctx context.Context, repo string, opts scm.ListOptions) ([]*scm.Hook, *scm.Response, error) {
 	path := fmt.Sprintf("api/v1/repos/%s/hooks?%s", repo, encodeListOptions(opts))
 	out := []*hook{}
@@ -148,6 +152,7 @@ type (
 		CreatedAt     time.Time `json:"created_at"`
 		UpdatedAt     time.Time `json:"updated_at"`
 		Permissions   perm      `json:"permissions"`
+		Archived      bool      `json:"archived"`
 	}
 
 	// gitea permissions details.
@@ -215,6 +220,7 @@ func convertRepository(src *repository) *scm.Repository {
 		Clone:     src.CloneURL,
 		CloneSSH:  src.SSHURL,
 		Link:      src.HTMLURL,
+		Archived:  src.Archived,
 	}
 }