+++ scm/driver/gitlab/repo.go
@@ -22,6 +22,7 @@ type repository struct {
 	PathNamespace string      `json:"path_with_namespace"`
 	DefaultBranch string      `json:"default_branch"`
 	Visibility    string      `json:"visibility"`
+	Archived      bool        `json:"archived"`
 	WebURL        string      `json:"web_url"`
 	SSHURL        string      `json:"ssh_url_to_repo"`
 	HTTPURL       string      `json:"http_url_to_repo"`
@@ -91,6 +92,10 @@ func (s *repositoryService) List(ctx context.Context, opts scm.ListOptions) ([]*
 	out := []*repository{}
 	res, err := s.client.do(ctx, "GET", path, nil, &out)
 	return convertRepositoryList(out), res, err
+}
+
+func (s *repositoryService) List2(ctx context.Context, orgSlug string, opts scm.ListOptions) ([]*scm.Repository, *scm.Response, error) {
+	return nil, nil, scm.ErrNotSupported
 }
 
 func (s *repositoryService) ListHooks(ctx context.Context, repo string, opts scm.ListOptions) ([]*scm.Hook, *scm.Response, error) {
@@ -180,6 +185,7 @@ func convertRepository(from *repository) *scm.Repository {
 		Namespace:  from.Namespace.Path,
 		Name:       from.Path,
 		Branch:     from.DefaultBranch,
+		Archived:   from.Archived,
 		Private:    convertPrivate(from.Visibility),
 		Visibility: convertVisibility(from.Visibility),
 		Clone:      from.HTTPURL,