+++ scm/driver/gogs/repo.go
@@ -45,6 +45,10 @@ func (s *repositoryService) List(ctx context.Context, _ scm.ListOptions) ([]*scm
 	return convertRepositoryList(out), res, err
 }
 
+func (s *repositoryService) List2(ctx context.Context, orgSlug string, opts scm.ListOptions) ([]*scm.Repository, *scm.Response, error) {
+	return nil, nil, scm.ErrNotSupported
+}
+
 func (s *repositoryService) ListHooks(ctx context.Context, repo string, _ scm.ListOptions) ([]*scm.Hook, *scm.Response, error) {
 	path := fmt.Sprintf("api/v1/repos/%s/hooks", repo)
 	out := []*hook{}